requirement.txt
python=3.8.10
torch=1.11.0+cu113
torchvision=0.12.0+cu113
pyequilib=0.5.4
typing-extensions=3.10.0.2
Pillow=9.1.0
opencv-python=4.5.5.64
tqdm=4.62.3
scikit-image=0.19.2
numpy=1.22.3
py360convert=0.1.0